SECURE FILE STORAGE ON CLOUD USING HYBRID CRYPTOGRAPHY/SECURE FILE STORAGE/SOURCE CODE/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
Flask==2.0.1
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycryptodome==3.10.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
requests==2.26.0
six==1.16.0
tinyec==0.4.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
